requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.0
attrs==25.4.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cssbeautifier==1.15.4
Django==6.0
django-debug-toolbar==6.2.0
django-filter==25.2
django-jazzmin==3.0.1
django-phonenumber-field==8.4.0
django-phonenumbers==1.0.1
djangorestframework==3.16.1
djlint==1.36.4
docutils==0.22.4
drf-spectacular==0.29.0
EditorConfig==0.17.1
flake8==7.3.0
flake8-docstrings==1.7.0
gunicorn==25.1.0
inflection==0.5.1
jsbeautifier==1.15.4
json5==0.12.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mccabe==0.7.0
packaging==26.0
pathspec==0.12.1
phonenumbers==9.0.22
pillow==12.1.0
pycodestyle==2.14.0
pycparser==2.23
pydocstyle==6.3.0
pyflakes==3.4.0
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
rpds-py==0.30.0
six==1.17.0
snowballstemmer==3.0.1
sqlparse==0.5.4
tqdm==4.67.1
typing_extensions==4.15.0
uritemplate==4.2.0
